requirements.txt
pydantic
pydantic-settings
fastapi
uvicorn
pyrebase4
firebase_admin